backend/requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.1.0
billiard==4.2.0
celery==5.3.5
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.0
Django==4.1.13
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-timezone-field==6.0.1
djangorestframework==3.14.0
drf-spectacular==0.26.5
flake8==6.1.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.19.2
jsonschema-specifications==2023.11.1
kombu==5.3.3
mccabe==0.7.0
packaging==23.2
pluggy==1.3.0
prompt-toolkit==3.0.41
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
pytest==7.4.3
pytest-django==4.7.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.31.0
requests==2.31.0
rpds-py==0.12.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.10
